refactor: extract shared PKCE utility
