+++ Tests/Domains/Auth/HostedAuthTests.swift
@@ -32,17 +32,17 @@ struct HostedAuthProtocolTests {
   func pkceChallengeMatchesRFC7636Vector() {
     let verifier = "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
 
-    #expect(HostedAuthPKCE.challenge(for: verifier) == "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM")
+    #expect(PKCE.challenge(for: verifier) == "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM")
   }
 
   @Test
   func generatedPKCEPairUsesS256CompatibleValues() throws {
-    let pair = try HostedAuthPKCE.generatePair()
+    let pair = try PKCE.generatePair()
 
     #expect(pair.verifier.count == 43)
     #expect(pair.challenge.count == 43)
     #expect(pair.verifier.allSatisfy { $0.isLetter || $0.isNumber || $0 == "-" || $0 == "_" })
-    #expect(pair.challenge == HostedAuthPKCE.challenge(for: pair.verifier))
+    #expect(pair.challenge == PKCE.challenge(for: pair.verifier))
   }
 
   @Test
@@ -145,7 +145,7 @@ struct HostedAuthFlowTests {
         guard let createParams = createParams.value else {
           throw ClerkClientError(message: "Missing create params in test.")
         }
-        #expect(HostedAuthPKCE.challenge(for: params.codeVerifier) == createParams.codeChallenge)
+        #expect(PKCE.challenge(for: params.codeVerifier) == createParams.codeChallenge)
         return ClientServiceResponse(client: redeemedClient, requestSequence: nil, serverDate: nil)
       }
     )
