+++ Tests/Domains/Auth/SignIn/SignInTests.swift
@@ -170,7 +170,7 @@ struct SignInTests {
     #expect(params.1.strategy == .emailLink)
     #expect(params.1.emailAddressId == "ema_123")
     #expect(params.1.redirectUri == Clerk.shared.options.redirectConfig.redirectUrl)
-    #expect(params.1.codeChallengeMethod == MagicLinkPKCE.codeChallengeMethod)
+    #expect(params.1.codeChallengeMethod == PKCE.codeChallengeMethod)
     #expect(params.1.codeChallenge?.isEmpty == false)
 
     let pendingFlow = try #require(magicLinkStore.load())
